+++ src/Atata.WebDriverExtras/Exceptions/ElementExceptionFactory.cs
@@ -3,21 +3,13 @@
 // TODO: v4. Remove ElementExceptionFactory. Move methods to ElementNotFoundException and ElementNotMissingException.
 public static class ElementExceptionFactory
 {
+    [Obsolete("Instead use ElementNotFoundException.Create(...)")] // Obsolete since v4.0.0.
     public static ElementNotFoundException CreateForNotFound(string? elementName = null, By? by = null, ISearchContext? searchContext = null) =>
-        CreateForNotFound(
-            new SearchFailureData
-            {
-                ElementName = elementName,
-                By = by,
-                SearchContext = searchContext
-            });
+        ElementNotFoundException.Create(elementName, by, searchContext);
 
-    public static ElementNotFoundException CreateForNotFound(SearchFailureData searchFailureData)
-    {
-        string message = (searchFailureData ?? new SearchFailureData()).ToStringForElementNotFound();
-
-        return new ElementNotFoundException(message);
-    }
+    [Obsolete("Instead use ElementNotFoundException.Create(...)")] // Obsolete since v4.0.0.
+    public static ElementNotFoundException CreateForNotFound(SearchFailureData searchFailureData) =>
+        ElementNotFoundException.Create(searchFailureData);
 
     public static ElementNotMissingException CreateForNotMissing(string? elementName = null, By? by = null, ISearchContext? searchContext = null) =>
         CreateForNotMissing(